+++ lib/domain/behaviours/reports/report.ex
@@ -1,0 +1,3 @@
+defmodule DistributedPerformanceAnalyzer.Domain.Behaviours.Reports.Report do
+  # @callback csv_save
+end
